Chap1 Helloworld
In this program I learned that the print function, "print()" outputs, to the terminal, whatever is inside the parenthesis, most commonly a string. I also learned how to run a program, you use cd to navigate to the directory in which your program is located and then you do python3 filename to run the program. In the next chapter I will be learning about data types.