2.1 Exercise: variable assignment
In this program I learned how to assign a string of characters ("hello to everyone seeing this") to a variable "message." Now whenever that variable is placed in a print function like this, print(var_name) it will print whatever the variable was assigned to. I can see how this will come in handy for larger programs that have to be changed dynamically and efficiently.