Exercise 3.1: Name
In this program I printed out a list of my friends names using the list feature in python. I also learned the def main() aspect from CS50P which told me that in python you can create your own functions. I also learned that you can use the string manipulating functions like .title() on list elements that are strings. I used a variety of indexes like 1, 2, -2, and -1 to internalize the fact that -1 index is the last element on the list and -2 is the second to last element on the list, and so on so forth.