Exercise 3.1: Name
In this program i learned to use the list function in python to create a list of names. I also knew, from CS50P,  that to create a function you can use the def statement so I used that to create my main function. I also learned that you can use sting manipulating method .title() on elements of a list if they are of string type.