Exercise 3.8: Seeing the World
This program was just to familiarize myself with organizing lists. However I was stumped for a while because the last two blocks of code returned "none" but then I figured out that I couldn't use f-strings like the others because .reverse() and .sort() permanently changes the list so I have to initialize the change and then call the original list.