Exercise 4.1: Pizzas
In this lesson I learned how to use a for loop in a list. I learned that indentation is vital to get a program to print out exactly how you want it because if I were to have indented the last print call it would have printed "Pizza is one of my favorite foods" for every single iteration of element on the list. I also learned that you should have good naming conventions for for loops like for "dog in dogs" or "for cat in cats" so it will be easier to read and keep track of in the future.