Exercise 4.4: One Million
PROMPT: "Make a list of the numbers from one to one million, and then use a for loop to print the numbers." In this program I implemented a list comprehension to familiarize myself with the syntax. I printed out value for value in range (1, 1000001) to print out a list of numbers from 1 to one million.