Exercise 4.6: Odd Numbers
In this programmed I practiced passing a 3rd value into the range() function which essentially increments the value of the 3rd value starting from the first value. In this program I wanted to print the odd numbers from 1 - 20 and since our starting value was 1 we can just increment by 2 to make it perpetually increment to an odd number until it hits at 20 or past it.