Exercise 5.9: Checking Usernames
I implemented a program that checks if a username has already been taken. I used two lists and compared them together to see if a certain username has already been taken, I achieved this through a for loop and if the new user.lower() is in the current user list we have to say it has been taken, else welcome them to the platform.