Exercise 4, week 2: Plates.py
In this program I had to make a validity test for plates with many conditions. This was very hard to make because of the numerous conditions, yet I prevailed. I had the most trouble with the conditions of the first digit cannot be 0 and that numbers can only be towards the end.