Exercise 8.11: Archived Messages
Just like the previous exercise but prevented the original message list form disappearing by using a clone (created by splitting the list) to pass as an argument instead of the real list. I also acknowledge that they are not exact clones as the messages print in reverse order but I can easily fix that by doing messages.reverse().