exercise-9.8: privileges
In this program I created a new class called Privileges so we can add more functions pertaining to privileges without cluttering the admin class. In this case we moved the show_privileges() method to the class. In the future we can add other methods such as remove_privilege(), add_privilege(), and etc.