exercise-1: fuel
This took me a while to make because I was quite stuck on how to separate the input into numerator and denominator but I found out that the string method .seperate() separates the while string with the center being what you pass through it. With this I got over that bump. However, It also took me a while to meet all the error conditions with a lot of them being ValueErrors and making sure to raise them and except them. I also got a little stuck with 2/3 evaluating as 66.6% instead of 66.7% but I solved that by using the round() function.