exercise-2: taqueria
This program was easier to make, the one part I got stuck on was the EOFError part because my computer was pasting the directory name instead of stopping the inputs, but that was just my side. I did the test cases and passed the EOFError part indicating that it was just an error on my side and not my program. Overall this was pretty easy to implement, but I had to search up how to format the numbers so that it was in two decimal places and found out you can achieve this through using an f-string then providing the :.2f format extension to your variable.