exercise-1: lines
I created a program that counts how many lines of code there is in a file, excluding comments and blank lines. The biggest problem I ran into was not including the blank lines, however i fixed that by adding a second condition to my if statement. I added the if lines which essentially checks if the line has anything in it.