exercise-3: scourgify
This program was pretty fun to implement. I learned a lot with reading and writing from and to files. The only trouble I had was getting the correct format: however, I solved this problem by reviewing the shorts about reading/writing CSV files and remembered that I could import the CSV library to use the DictReader and DictWriter methods to read from the before file and write to the after file.