exercise-4: shirt
This program was pretty difficult to implement, I had to study the methods/functions that the PIL library had to offer. The hardest part was dealing with the  formatting and debugging. The easiest part for me was creating the checks to see if the requirements for the sys arguments were correct.