exercise-2: watch
I am still finding regex really hard to understand because of its cryptic style. However, I did learn something new about greedy and non-greedy applications. For example: the segment ".*?" takes in any characters except new lines with 0 or more repetitions until it finds its FIRST match, while ".*" is greedy and does this until its very last match. This knowledge helped me solve this problem as the last test I needed to pass was to make sure that the src link was inside the iframe. To do this I had to essentially find the first instance of src with <iframe in the beginning while tuning out the excess information.