exercise-3: working
This program took me a day or two to implement, the hardest part was dealing with all the constraints of the 24 hour time. This program works, however, I believe this is suboptimal. I will take some time to research what I can do to make this program less cluttered and more optimal.