exercise-4: um
This program was pretty straightforward to implement. The only thing I had trouble with was making sure it was only counting the word um, not the string of characters "um" inside of words. I read through the documentation and found that \b matches words that starts and ends with that string of words, thus creating a sort of word boundary.