exercise-1: seasons
This program that I created calculates the delta of minutes from the user's inputted date of birth and today's date. I then use the library inflect to transform those said words into more readable words. The only thing I was having trouble with was the creating the inflect engine as i had forgotten to put the parentheses in the p = inflect.engine. Which had great implications as I was not able to use the number_to_words method.